audio system (CORI_81, CORI_65, CORI_66). new error handling (CORI_70)
audio system uses sdl3_mixer under the hood. new way of handling errors is designed to be used with std::expected but CoriError can be used as a regular exception as well, its derived from std::exception. PossibleErrors class allows to use std::variant with CoriError and std::expected, this allows to specify multiple possible CoriError specializations as an error in one std::expected. The logic is done, only thing left is to adopt this new error handling everywhere in the engine.

+++ CoriEngine/src/Renderer/OpenGL/GL_VertexArray.cpp
@@ -72,7 +72,7 @@ namespace Cori {
 		const VBLayout& layout = vertexBuffer->GetLayout();
 		for (const auto& element : layout) {
 			if (element.m_Type == ShaderDataType::Mat3) {
-				for (int i = 0; i < element.GetComponentCount(); i++) {
+				for (size_t i = 0; i < element.GetComponentCount(); i++) {
 					glEnableVertexAttribArray(index);
 					glVertexAttribPointer(index, static_cast<GLint>(element.GetComponentCount()), ShaderDataTypeToGLDataType(element.m_Type), element.m_Normalized ? GL_TRUE : GL_FALSE, static_cast<GLsizei>(layout.GetStride()), reinterpret_cast<const void*>(element.m_Offset + sizeof(float) * element.GetComponentCount() * i));
